# agent/src/agent/generated/api/main.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-08-25T13:06:54+00:00

from __future__ import annotations

from typing import Optional, Union

from fastapi import FastAPI
from pydantic import constr

from .models import (
    AgentInput,
    AgentListResponse,
    AgentName,
    AgentNameResponse,
    AgentResponse,
    AgentUsageResponse,
    ErrorDetail,
    HTTPValidationError,
    ProbeError,
    ProbeOk,
    ReferenceListResponse,
)

app = FastAPI(
    title='Agent',
    description='Agent registry service',
    version='0.1.0',
    servers=[{'url': 'http://localhost:8003', 'description': 'Local agent service'}],
)


@app.get('/api/v1/agents', response_model=AgentListResponse)
def list_agents() -> AgentListResponse:
    """
    List all agents
    """
    pass


@app.get(
    '/api/v1/agents/usage/knowledge/{knowledge_name}',
    response_model=ReferenceListResponse,
)
def list_knowledge_usage(knowledge_name: constr(min_length=1)) -> ReferenceListResponse:
    """
    List agents that reference a knowledge source
    """
    pass


@app.get('/api/v1/agents/usage/mcp/{mcp_name}', response_model=ReferenceListResponse)
def list_mcp_usage(mcp_name: constr(min_length=1)) -> ReferenceListResponse:
    """
    List agents that reference an MCP server
    """
    pass


@app.get(
    '/api/v1/agents/usage/skill/{skill_name}', response_model=ReferenceListResponse
)
def list_skill_usage(skill_name: constr(min_length=1)) -> ReferenceListResponse:
    """
    List agents that reference a skill
    """
    pass


@app.get(
    '/api/v1/agents/{name}',
    response_model=AgentResponse,
    responses={'404': {'model': ErrorDetail}, '422': {'model': HTTPValidationError}},
)
def get_agent(
    name: AgentName,
) -> Union[AgentResponse, ErrorDetail, HTTPValidationError]:
    """
    Get an agent by name
    """
    pass


@app.post(
    '/api/v1/agents/{name}',
    response_model=AgentNameResponse,
    status_code=201,
    responses={'409': {'model': ErrorDetail}, '422': {'model': HTTPValidationError}},
)
def create_agent(
    name: AgentName, body: AgentInput = ...
) -> Union[AgentNameResponse, ErrorDetail, HTTPValidationError]:
    """
    Create an agent
    """
    pass


@app.put(
    '/api/v1/agents/{name}',
    response_model=AgentNameResponse,
    responses={'404': {'model': ErrorDetail}, '422': {'model': HTTPValidationError}},
)
def update_agent(
    name: AgentName, body: AgentInput = ...
) -> Union[AgentNameResponse, ErrorDetail, HTTPValidationError]:
    """
    Replace an agent
    """
    pass


@app.delete(
    '/api/v1/agents/{name}',
    response_model=None,
    status_code=204,
    responses={
        '404': {'model': ErrorDetail},
        '422': {'model': HTTPValidationError},
        '409': {'model': ErrorDetail},
        '502': {'model': ErrorDetail},
    },
)
def delete_agent(name: AgentName) -> Optional[Union[ErrorDetail, HTTPValidationError]]:
    """
    Delete an agent
    """
    pass


@app.get(
    '/api/v1/agents/{name}/usage',
    response_model=AgentUsageResponse,
    responses={
        '404': {'model': ErrorDetail},
        '422': {'model': HTTPValidationError},
        '502': {'model': ErrorDetail},
    },
)
def get_agent_usage(
    name: AgentName,
) -> Union[AgentUsageResponse, ErrorDetail, HTTPValidationError]:
    """
    List crews that use an agent
    """
    pass


@app.get('/probe', response_model=ProbeOk, responses={'503': {'model': ProbeError}})
def probe() -> Union[ProbeOk, ProbeError]:
    """
    Check service and database health
    """
    pass