# crew/src/crew/generated/api/models.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-08-25T12:47:34+00:00

from __future__ import annotations

from typing import List, Literal, Optional, Union

from pydantic import AwareDatetime, BaseModel, ConfigDict, Field, RootModel, constr


class CrewName(
    RootModel[constr(pattern=r'^[a-zA-Z0-9_-]{2,50}$', min_length=2, max_length=50)]
):
    root: constr(pattern=r'^[a-zA-Z0-9_-]{2,50}$', min_length=2, max_length=50) = Field(
        ..., examples=['travel-assistant']
    )


class CrewInput(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    model: constr(min_length=1)
    role: constr(min_length=1)
    goal: constr(min_length=1)
    backstory: constr(min_length=1)
    agents: Optional[List[str]] = []


class CrewResponse(BaseModel):
    name: CrewName
    model: str
    role: str
    goal: str
    backstory: str
    agents: List[str]
    updated_at: AwareDatetime


class CrewListResponse(BaseModel):
    crews: List[CrewResponse]


class CrewNameResponse(BaseModel):
    name: CrewName


class ReferenceListResponse(BaseModel):
    references: List[CrewName] = Field(..., examples=[['travel-assistant']])


class ProbeOk(BaseModel):
    service: Literal['ok']
    database: Literal['ok']


class ProbeError(BaseModel):
    error: str


class ErrorDetail(BaseModel):
    detail: str


class ValidationError(BaseModel):
    loc: Optional[List[Union[str, int]]] = None
    msg: Optional[str] = None
    type: Optional[str] = None


class HTTPValidationError(BaseModel):
    detail: Optional[List[ValidationError]] = None