# knowledge/src/knowledge/generated/api/models.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-08-25T12:26:01+00:00

from __future__ import annotations

from typing import List, Literal, Optional, Union

from pydantic import AwareDatetime, BaseModel, ConfigDict, Field, RootModel, constr


class KnowledgeName(
    RootModel[constr(pattern=r'^[a-zA-Z0-9_-]{2,50}$', min_length=2, max_length=50)]
):
    root: constr(pattern=r'^[a-zA-Z0-9_-]{2,50}$', min_length=2, max_length=50) = Field(
        ..., examples=['cities']
    )


class KnowledgeInput(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    content: constr(min_length=1, max_length=1000000)


class KnowledgeResponse(BaseModel):
    name: KnowledgeName
    content: str
    updated_at: AwareDatetime


class KnowledgeListResponse(BaseModel):
    knowledge: List[KnowledgeResponse]


class KnowledgeNameResponse(BaseModel):
    name: KnowledgeName


class KnowledgeUsageResponse(BaseModel):
    agents: List[str] = Field(..., examples=[['researcher']])
    skills: List[str] = Field(..., examples=[['weather-lookup']])


class ProbeOk(BaseModel):
    service: Literal['ok']
    database: Literal['ok']


class ProbeError(BaseModel):
    error: str


class ErrorDetail(BaseModel):
    detail: str


class ValidationError(BaseModel):
    loc: Optional[List[Union[str, int]]] = None
    msg: Optional[str] = None
    type: Optional[str] = None


class HTTPValidationError(BaseModel):
    detail: Optional[List[ValidationError]] = None