# mcp/src/mcp_registry/generated/api/models.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-08-25T13:19:17+00:00

from __future__ import annotations

from typing import Any, Dict, List, Literal, Optional, Union

from pydantic import BaseModel, ConfigDict, Field, RootModel, constr


class McpName(
    RootModel[constr(pattern=r'^[a-zA-Z0-9_-]{2,50}$', min_length=2, max_length=50)]
):
    root: constr(pattern=r'^[a-zA-Z0-9_-]{2,50}$', min_length=2, max_length=50) = Field(
        ..., examples=['book-flight']
    )


class McpInput(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    url: constr(min_length=1)
    headers: Optional[Dict[str, str]] = None


class McpServerConfig(BaseModel):
    url: str
    headers: Optional[Dict[str, str]] = None


class McpListResponse(BaseModel):
    mcpServers: Dict[str, McpServerConfig] = Field(
        ...,
        examples=[
            {
                'book-flight': {
                    'url': 'https://book-flight-mcp.demo.abdollahpour.com/mcp'
                }
            }
        ],
    )


class McpNameResponse(BaseModel):
    name: McpName


class McpUsageResponse(BaseModel):
    agents: List[str] = Field(..., examples=[['researcher']])
    skills: List[str] = Field(..., examples=[['weather-lookup']])


class McpTool(BaseModel):
    name: str
    description: Optional[str] = None
    input_schema: Dict[str, Any]


class McpToolListResponse(BaseModel):
    tools: List[McpTool]


class CallToolInput(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    arguments: Optional[Dict[str, Any]] = {}


class CallToolResult(BaseModel):
    is_error: bool
    content: List[Dict[str, Any]]
    structured_content: Optional[Any] = None


class ProbeOk(BaseModel):
    service: Literal['ok']
    database: Literal['ok']


class ProbeError(BaseModel):
    error: str


class ErrorDetail(BaseModel):
    detail: str


class ValidationError(BaseModel):
    loc: Optional[List[Union[str, int]]] = None
    msg: Optional[str] = None
    type: Optional[str] = None


class HTTPValidationError(BaseModel):
    detail: Optional[List[ValidationError]] = None