# skill/src/skill/generated/api/models.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-08-24T11:20:30+00:00

from __future__ import annotations

from typing import List, Literal, Optional, Union

from pydantic import AwareDatetime, BaseModel, ConfigDict, Field, RootModel, constr


class SkillName(
    RootModel[constr(pattern=r'^[a-zA-Z0-9_-]{2,50}$', min_length=2, max_length=50)]
):
    root: constr(pattern=r'^[a-zA-Z0-9_-]{2,50}$', min_length=2, max_length=50) = Field(
        ..., examples=['example_skill']
    )


class SkillInput(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    description: constr(min_length=1, max_length=500)
    skill_md: constr(min_length=1, max_length=1000000)
    tools_required: Optional[List[str]] = Field(
        [], description='Must be names from GET /api/v1/skills/tools'
    )
    knowledge: Optional[List[str]] = []


class SkillResponse(BaseModel):
    name: SkillName
    description: str
    skill_md: str
    tools_required: List[str]
    knowledge: List[str]
    updated_at: AwareDatetime


class SkillListResponse(BaseModel):
    skills: List[SkillResponse]


class SkillNameResponse(BaseModel):
    name: SkillName


class ToolListResponse(BaseModel):
    tools: List[str] = Field(..., examples=[['CodeInterpreterTool', 'HttpCallTool']])


class ReferenceListResponse(BaseModel):
    references: List[SkillName] = Field(..., examples=[['example_skill']])


class SkillUsageResponse(BaseModel):
    agents: List[str] = Field(..., examples=[['researcher']])


class ProbeOk(BaseModel):
    service: Literal['ok']
    database: Literal['ok']


class ProbeError(BaseModel):
    error: str


class ErrorDetail(BaseModel):
    detail: str


class ValidationError(BaseModel):
    loc: Optional[List[Union[str, int]]] = None
    msg: Optional[str] = None
    type: Optional[str] = None


class HTTPValidationError(BaseModel):
    detail: Optional[List[ValidationError]] = None